feat(client): zombies, drops, crates and combat, host-simulated for every player
Entities (lib/entities, twins of entities/*.ts): the A* pathfinder over
standing cells, the ZombieManager (four variants with the same stats,
night-scaled spawn groups on grass/sand on the player's layer, chase on the
path or straight, attacks with cooldown, breaking the block in the way with
per-block hit points, dawn burn, death timer), drops that fall, settle and
are picked up when walked over, and loot crates that settle where a player
fell. Tests mirror entities/__tests__/{zombies,crates,drops}.test.ts.

Host sim (lib/game/host_sim.dart + avatar.dart, the host-only half of
Game.ts): the night schedule (groups over 70 % of the night, "Night N —
they are coming", dawn burn + "you survived"), zombie hits on the nearest
avatar with poison for Toxic, health regen after 8 s, sword swing (reach,
arc, knockback), rifle fire (voxel raycast, first zombie box on the ray,
headshots, Soldier resistance), reload from ammo over 2 s, death → the
inventory becomes a crate → respawn after 5 s at the bed/pad, loot via F,
placing props (bed with its partner cell), breaking with drops. Every
avatar — the host's and each joiner's — goes through the same rules; a
remote avatar's private state is queued and sent by HostSession, a client
sends its actions to the host and mirrors zombies/drops/crates from the
snapshots. The host streams the world around every avatar and never drops
a column a zombie or drop stands in.

Game: FrameInput gains primaryHeld (auto rifle), reload, dropHeld and
interact; the HUD shows ammo, the "N out there" counter, poison, the hurt
flash, the death screen with the respawn countdown and third person while
dead. Saves carry typed zombies/drops/crates and departed visitors' gear;
dawn posts the score and autosaves. Zombies, drops and crates render as
vertex-coloured boxes through the chunk pipeline until the glTF loader
lands. Debug builds skip to sunset on N.

Protocol: PlaceBlock.yaw and prop yaw are doubles (radians), as on the web.
Review fixes: a client finds itself in snapshots by welcome.you; a zombie's
hit only lands within attack range + 0.8 as in Game.ts.

+++ client/lib/world/world.dart
@@ -31,8 +31,8 @@ final class PropMeta {
   final int y;
   final int z;
 
-  /// facing, multiples of π/2
-  final int yaw;
+  /// facing in radians, a multiple of π/2
+  final double yaw;
 
   /// for multi-cell props: the other cell; `primary` cells own the model
   final (int, int, int)? partner;
